requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bitsandbytes==0.40.0.post4
black==23.7.0
bleach==6.0.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools==5.3.0
Cerberus==1.3.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cloudpickle==2.2.1
cmake==3.26.3
comm==0.1.2
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cubinlinker-cu11==0.3.0.post1
cuda-python==11.8.1
cudf-cu11==23.4.0.1681363056
cuml-cu11==23.4.0.1681368248
cupy-cuda11x==11.6.0
cycler==0.11.0
Cython==0.29.34
cytoolz==0.11.0
dask==2023.3.2
dask-cuda==23.4.0
dask-cudf-cu11==23.4.0.1681364034
datasets==2.10.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distributed==2023.3.2.1
docker-pycreds==0.4.0
entrypoints==0.3
evaluate==0.4.0
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.16.2
fastrlock==0.8.1
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
flake8==3.7.9
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
funcy==1.15
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gin-config==0.3.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.1
google-auth-oauthlib==1.0.0
gradio==3.36.1
gradio_client==0.2.9
grpcio==1.53.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==4.3.21
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
linkify-it-py==2.0.2
lit==16.0.1
littleutils==0.2.2
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
munch==2.5.0
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvtx==0.2.5
oauthlib==3.2.2
ogb==1.3.5
omegaconf==2.3.0
openai==0.27.8
orjson==3.9.2
outdated==0.2.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.1
pathtools==0.1.2
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.0.0
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptxcompiler-cu11==0.7.0.post1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.0.2
pydantic_core==2.1.2
pydub==0.25.1
pyflakes==2.1.1
Pygments==2.14.0
pylibraft-cu11==23.4.0.1681363053
pynvml==11.4.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyre-extensions==0.0.23
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.6
python-multipart==0.0.6
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==5.4
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
quinine @ git+https://github.com/krandiash/quinine.git@a820365e08ba1277d9246bbff7d143154d1ba8c4
raft-dask-cu11==23.4.0.1681367712
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rmm-cu11==23.4.0.1681362075
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.16.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.27.0
submitit==1.4.5
sympy==1.11.1
tabulate==0.9.0
tb-nightly==2.13.0a20230406
tblib==1.7.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
timm==0.6.12
tinycss2==1.2.1
tokenize-rt==5.1.0
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
toposort==1.5
torch==2.0.0+cu117
torchaudio==0.13.1
torchmetrics==0.10.3
torchvision==0.15.0+cu117
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@08667050223a6f27b8cd3a057bc06585ff190ecf
treelite==3.2.0
treelite-runtime==3.2.0
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.7.1
uc-micro-py==1.0.2
ucx-py-cu11==0.31.0.1681362077
unrar==0.4
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
uvicorn==0.22.0
wandb==0.13.10
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wilds==1.2.2
xformers==0.0.18
xxhash==3.2.0
yacs==0.1.8
yapf==0.29.0
yarl==1.8.2
zict==3.0.0
zipp==3.13.0
